Create Program1_5.c
The program prints a horizontal line of asterisks based on a given number. The Accept function takes an integer as input and uses a for loop to print that many * characters consecutively. In main, the number of asterisks is set to 5, and the Accept function is called to display the line.