Create Program14_5.c
This C program counts the number of digits in a given integer that are **less than 6**. It first converts any negative number to positive to simplify digit extraction. Then, using a `while` loop, it extracts each digit by taking the remainder when divided by 10. If the digit is less than 6, a counter is incremented. The process continues until all digits are processed, and finally, the program prints the total count of digits that are less than 6.

**Time Complexity:** O(n), where n is the number of digits in the input number.