Create Program15_1.c
This program counts the number of **even digits** in a given integer.

It works by extracting each digit of the number using the modulus operator `%` and checking if it is divisible by 2. If the digit is even, a counter is incremented. The process continues until all digits are processed.

Finally, it prints the total count of even digits.

**Time Complexity:** O(n), where n is the number of digits in the input number.